moved AppyxLogger into multiplatform utils
