+++ appyx-components/demos/common/src/commonMain/kotlin/com/bumble/appyx/components/demos/cards/ui/CardsMotionController.kt
@@ -10,7 +10,6 @@ import com.bumble.appyx.components.demos.cards.CardsModel
 import com.bumble.appyx.components.demos.cards.CardsModel.State.Card.InvisibleCard.VotedCard.VOTED_CARD_STATE.LIKED
 import com.bumble.appyx.components.demos.cards.operation.VoteLike
 import com.bumble.appyx.components.demos.cards.operation.VotePass
-import com.bumble.appyx.interactions.AppyxLogger
 import com.bumble.appyx.interactions.core.model.transition.Operation
 import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
 import com.bumble.appyx.interactions.core.ui.context.UiContext
@@ -22,6 +21,7 @@ import com.bumble.appyx.interactions.core.ui.property.impl.Scale
 import com.bumble.appyx.interactions.core.ui.property.impl.ZIndex
 import com.bumble.appyx.interactions.core.ui.state.MatchedTargetUiState
 import com.bumble.appyx.transitionmodel.BaseMotionController
+import com.bumble.appyx.utils.multiplatform.AppyxLogger
 
 class CardsMotionController<InteractionTarget : Any>(
     uiContext: UiContext,
