+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/model/transition/Keyframes.kt
@@ -1,6 +1,6 @@
 package com.bumble.appyx.interactions.core.model.transition
 
-import com.bumble.appyx.interactions.AppyxLogger
+import com.bumble.appyx.utils.multiplatform.AppyxLogger
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.distinctUntilChanged
