--- utils/multiplatform/src/androidMain/kotlin/com/bumble/appyx/utils/multiplatform/Logger.kt
@@ -1,9 +1,0 @@
-package com.bumble.appyx.utils.multiplatform
-
-import android.util.Log
-
-actual class Logger {
-    actual fun d(tag: String, message: String) {
-        Log.d(tag, message)
-    }
-}
