--- utils/multiplatform/src/commonMain/kotlin/com/bumble/appyx/utils/multiplatform/Logger.kt
@@ -1,5 +1,0 @@
-package com.bumble.appyx.utils.multiplatform
-
-expect class Logger() {
-    fun d(tag: String, message: String)
-}
