--- utils/multiplatform/src/jsMain/kotlin/com/bumble/appyx/utils/multiplatform/Logger.kt
@@ -1,7 +1,0 @@
-package com.bumble.appyx.utils.multiplatform
-
-actual class Logger {
-    actual fun d(tag: String, message: String) {
-        println("$tag: $message")
-    }
-}
